+++ .env.example
@@ -1,0 +1,2 @@
+GESTURE_AI_CONFIG_PATH = 'configs/gesture_config.json'
+GESTURE_AI_DEFAULT_CONFIG = 'configs/gesture_config_default.json'
