+++ configs/gesture_config_default.json
@@ -1,0 +1,38 @@
+{
+  "default_mode": {
+    "open_palm": "paste",
+    "fist": "open_terminal",
+    "thumbs_up": "open_browser",
+    "ok_sign": "lock_computer",
+    "thumbs_down": "open_file_explorer",
+    "three_fingers": "None",
+    "rock_on": "None",
+    "pinch": "pause_video",
+    "swipe_left": "swipe_left_tab",
+    "swipe_right": "swipe_right_tab"
+  },
+  "word_mode": {
+    "pointing_finger": "select_all",
+    "open_palm": "paste",
+    "peace_sign": "copy",
+    "fist": "cut",
+    "thumbs_up": "undo",
+    "thumbs_down": "redo",
+    "three_fingers": "None",
+    "rock_on": "None"
+  },
+  "media_mode": {
+    "swipe_left": "decrease_volume",
+    "swipe_right": "increase_volume",
+    "three_fingers": "None",
+    "rock_on": "None",
+    "pinch": "pause_video",
+    "open_palm": "pause_video"
+  },
+  "presentation_mode": {
+    "open_palm": "start_presentation",
+    "swipe_right": "next_slide",
+    "swipe_left": "previous_slide",
+    "three_fingers": "exit_presentation"
+  }
+}
